PEZY-418: Backend | Driver | Create driverService.js: getDriverByLicense()
